pyproject.toml
[project]
name = "final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "modal>=1.4.1",
    "numpy>=2.4.4",
    "python-dotenv>=1.2.2",
]
